:art: rename to match function name
